machine-learning/pyproject.toml
[project]
name = "immich-ml"
version = "2.7.5"
description = ""
authors = [{ name = "Hau Tran", email = "alex.tran1502@gmail.com" }]
requires-python = ">=3.11,<4.0"
readme = "README.md"
dependencies = [
    "aiocache>=0.12.1,<1.0",
    "fastapi>=0.95.2,<1.0",
    "gunicorn>=21.1.0",
    "huggingface-hub>=0.20.1,<1.0",
    "insightface>=0.7.3,<1.0",
    "numpy<2.4.0",
    "opencv-python-headless>=4.7.0.72,<5.0",
    "orjson>=3.9.5",
    "pillow>=12.2,<12.3",
    "pydantic>=2.0.0,<3",
    "pydantic-settings>=2.5.2,<3",
    "python-multipart>=0.0.6,<1.0",
    "rich>=13.4.2",
    "tokenizers>=0.15.0,<1.0",
    "uvicorn[standard]>=0.22.0,<1.0",
    "rapidocr>=3.1.0",
]

[dependency-groups]
test = [
    "httpx>=0.24.1",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
]
types = [
    "types-pyyaml>=6.0.12.20241230",
    "types-requests>=2.32.0.20250306",
    "types-setuptools>=75.8.2.20250305",
    "types-simplejson>=3.20.0.20250218",
    "types-ujson>=5.10.0.20240515",
]
lint = [
    "mypy>=1.3.0",
    "ruff>=0.0.272",
    { include-group = "types" },
]
dev = ["locust>=2.15.1", { include-group = "test" }, { include-group = "lint" }]

[project.optional-dependencies]
cpu = ["onnxruntime>=1.23.2,<2"]
cuda = ["onnxruntime-gpu>=1.23.2,<2"]
openvino = ["onnxruntime-openvino>=1.24.1,<2"]
armnn = ["onnxruntime>=1.23.2,<2"]
rknn = ["onnxruntime>=1.23.2,<2", "rknn-toolkit-lite2>=2.3.0,<3"]
rocm = ["onnxruntime-migraphx>=1.23.2,<2"]
segmentation = ["torch>=2.0", "mobile-sam"]

[tool.uv]
compile-bytecode = true

[tool.hatch.build.targets.sdist]
include = ["immich_ml"]

[tool.hatch.build.targets.wheel]
include = ["immich_ml"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
plugins = "pydantic.mypy"
follow_imports = "silent"
warn_redundant_casts = true
disallow_any_generics = true
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
per-file-ignores = { "test_main.py" = ["F403"] }

[tool.pytest.ini_options]
markers = ["providers", "ov_device_ids"]
